switch to the new estimator
