+++ runtime/runtime-params-estimator/Cargo.toml
@@ -41,7 +41,6 @@ glob = "0.3.0"
 walrus = "0.18.0"
 hex = "0.4"
 cfg-if = "1"
-once_cell = "1"
 num-traits = "0.2.12"
 libc = "0.2.81"
 wat = "1.0"
